Integrated the indexeddb with the resource

+++ scripts/services/ea-resource.js
@@ -194,7 +194,6 @@ angular.module('employeeApp.services')
                             throw "Expected between 0-4 arguments [params, data, success, error], got " +
                                 arguments.length + " arguments.";
                     }
-                    //dump(params);
                     /*
                      * RESETTING AREA:
                      * 
@@ -220,37 +219,67 @@ angular.module('employeeApp.services')
                     }
                     
                     /*
-                     * Preload data from the storage if the action method is
-                     * a GET call
+                     * Preloads Data from the indexedDB
+                     * 
+                     * We first must check if this is a request for data. 
+                     * After we check if the db is ready. Depending on whether 
+                     * it is ready or not we proceed in two different ways:
+                     * 
+                     * -db is ready:
+                     *     -check if to make query or get call
+                     * -db is not ready:
+                     *     -attach an on ready call and make the 
+                     *      appropriate get or query call
                      */
                      if (action.method == "GET" && !this.$$last_checked) {
-                         if (action.isArray) {
-                             db.query(function(data){
-                                 console.log(data);
-                                 for(var key in data){
-                                     //Loop for existing item in value
-                                     var index = indexOfId(value, data[key].id);
-                                     if (index != -1) {
-                                         /*
-                                          * In order not to waste resource we
-                                          * first check if the two items are equal or not.
-                                          * If they are not equal then we perform an extend
-                                          */
-                                         if (!angular.equals(value[index], data[key])) {
-                                             angular.extend(value[index], data[key]);
+                         if(db.ready){
+                             if (action.isArray) {
+                                 db.query(function(data){
+                                     $rootScope.safeApply(function(){
+                                         for(var key in data){
+                                             //Loop for existing item in value
+                                             var index = indexOfId(value, data[key].id);
+                                             if (index != -1) {
+                                                 angular.copy(data[key], value[index]);
+                                             } else {
+                                                 value.push(new Resource(data[key]));
+                                             }
                                          }
-                                         
-                                     } else {
-                                         //Add new resource
-                                         value.push(new Resource(data[key]));
-                                     }
+                                     });
+                                 });
+                             } else {
+                                 db.get(params.id, function(response){
+                                     $rootScope.safeApply(function(){
+                                        angular.extend(value, new Resource(response));
+                                     });
+                                 });
+                             }
+                         }else{
+                             if (action.isArray) {
+                                 db.onready = function(){
+                                     db.query(function(data){
+                                         $rootScope.safeApply(function(){
+                                             for(var key in data){
+                                                 //Loop for existing item in value
+                                                 var index = indexOfId(value, data[key].id);
+                                                 if (index != -1) {
+                                                     angular.copy(data[key], value[index]);
+                                                 } else {
+                                                     value.push(new Resource(data[key]));
+                                                 }
+                                             }
+                                         });
+                                     });
                                  }
-                                 
-                             });
-                             
-                         } else {
-                             //Set Reference as new Resource
-                             value = new Resource(storage.get(params.id));
+                             } else {
+                                 db.onready = function(){
+                                     db.get(params.id, function(response){
+                                         $rootScope.safeApply(function(){
+                                            angular.extend(value, new Resource(response));
+                                         });
+                                     });
+                                 }
+                             }
                          }
                      }
                      
@@ -324,9 +353,12 @@ angular.module('employeeApp.services')
                          * both return responses we would save this to the storage. 
                          * For delete requests, we would have to delete the item
                          */
-                      
-                        action.method == "DELETE" ? db.remove(params) : hasBody ? db.save(this) : db.save(value);  
                         
+                        if(db.ready){
+                            action.method == "DELETE" ? db.remove(params) : hasBody ? db.save(this) : db.save(value);  
+                        }
+                        
+                        //action.method == "DELETE" ? storage.remove(params) : hasBody ? storage.save(this) : storage.save(value);  
                         /*
                          * Last checked
                          */
